requirements.txt
asgiref==3.8.1
Django==4.2.11
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
isoweek==1.3.3
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
xlwt==1.3.0
